requirements.txt
google-generativeai 
flask
uvicorn 
faiss-cpu 
sentence-transformers 
requests
websockets==9.1
python-dotenv
nltk
spacy
